services/blockchain-audit-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
web3==6.10.0
python-dotenv==1.0.0
pydantic==2.3.0
sqlite3-api
